refactor: 💡 Configuracao Babel, react navigation e type script
Foram ajustadas um monte de coisas, quase desisti de raiva do babel com
o react Navigation
