--- .babelrc
@@ -1,11 +1,0 @@
-{
-  "plugins": [
-    ["module-resolver", {
-      "root": ["./src"],
-      "alias": {
-        "test": "./test",
-        "underscore": "lodash"
-      }
-    }]
-  ]
-}
