+++ .prettierrc.js
@@ -2,6 +2,5 @@ module.exports = {
   bracketSpacing: false,
   jsxBracketSameLine: true,
   singleQuote: true,
-  trailingComma: 'all',
-  "parser": 'typescript'
+  trailingComma: 'all'
 };
