+++ babel.config.js
@@ -1,26 +1,3 @@
 module.exports = {
-    presets: ['module:metro-react-native-babel-preset'],
-    plugins: [
-      'module-resolver',
-      {
-        root: ['./src'],
-        extensions: [
-          '.ios.ts',
-          '.android.ts',
-          '.ts',
-          '.ios.tsx',
-          '.android.tsx',
-          '.tsx',
-          '.jsx',
-          '.js',
-          '.json',
-        ],
-        alias: {
-          '@routes': 'UnFlix/src/routes',
-          '@components': 'UnFlix/src/components',
-        },
-      },
-    ],
-  };
-  
-  
+  presets: ['module:metro-react-native-babel-preset'],
+};
