+++ package.json
@@ -40,11 +40,9 @@
     "@types/react": "^16.9.56",
     "@types/react-native": "^0.63.33",
     "@types/react-test-renderer": "^16.9.3",
+    "@typescript-eslint/parser": "^4.7.0",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^25.1.0",
-    "babel-plugin-module-resolver": "^4.0.0",
-    "babel-plugin-root-import": "^6.6.0",
-    "babel-preset-react": "^6.24.1",
     "eslint": "^7.6.0",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-config-prettier": "^6.11.0",
