+++ src/screens/MovieDetail/MovieDetail.styles.ts
