+++ src/screens/MovieDetail/index.tsx
@@ -34,7 +34,7 @@ const MoviewDetail = ({ route }) => {
             <View style={{ flex: 1, justifyContent: 'center' }}>
               <Text style={Styles.textBody}>{movieDetail.overview}</Text>
               <Text style={Styles.title}>Lançamento</Text>
-              <Text style={Styles.text}>{movieDetail.release_date}</Text>
+              <Text style={Styles.textBody}>{movieDetail.release_date}</Text>
               <Text style={Styles.title}>Título original</Text>
               <Text style={Styles.textBody}>{movieDetail.original_title}</Text>
             </View>
