+++ src/screens/Trends/Trends.styles.ts
