+++ src/screens/Trends/index.tsx
